+++ src/pages/keyblocks/KeyblockMicroblocksPanel.vue
@@ -3,50 +3,46 @@
     <paginated-content
       v-model:page-index="pageIndex"
       pagination-style="history"
-      :entities="keyblockMicroblocks"
-      :total-count="microblocksCount"
+      :entities="microblocks"
+      :total-count="keyblockDetails.micro_blocks_count"
       :limit="limit"
       @prev-clicked="loadPrevMicroblocks"
       @next-clicked="loadNextMicroblocks">
       <keyblock-microblocks-table
-        v-if="keyblockMicroblocks"
-        :microblocks="keyblockMicroblocks"
+        v-if="microblocks"
+        :microblocks="microblocks"
         class="keyblock-microblock-panel__table"/>
       <keyblock-microblocks-table-condensed
-        v-if="keyblockMicroblocks"
-        :microblocks="keyblockMicroblocks"
+        v-if="microblocks"
+        :microblocks="microblocks"
         class="keyblock-microblock-panel__table-condensed"/>
     </paginated-content>
   </app-panel>
 </template>
 <script setup>
 import { storeToRefs } from 'pinia'
-import { useKeyblockDetailsStore } from '@/stores/keyblockDetailsStore'
-import KeyblockMicroblocksTable from '@/pages/keyblocks/KeyblockMicroblocksTable'
+import { computed } from 'vue'
+import { useKeyblockDetails } from '@/stores/keyblockDetails'
 import PaginatedContent from '@/components/PaginatedContent'
-import KeyblockMicroblocksTableCondensed from '~/pages/keyblocks/KeyblockMicroblocksTableCondensed'
+import KeyblockMicroblocksTable from '@/pages/keyblocks/KeyblockMicroblocksTable'
+import KeyblockMicroblocksTableCondensed from '@/pages/keyblocks/KeyblockMicroblocksTableCondensed'
+import { isDesktop } from '~/utils/screen'
 
-defineProps({
-  microblocksCount: {
-    type: Object,
-    required: true,
-  },
-})
-
-const { keyblockMicroblocks } = storeToRefs(useKeyblockDetailsStore())
-const { fetchKeyblockMicroblocks } = useKeyblockDetailsStore()
+const { keyblockMicroblocks: microblocks, keyblockDetails } = storeToRefs(useKeyblockDetails())
+const { fetchKeyblockMicroblocks } = useKeyblockDetails()
 const route = useRoute()
-
-const limit = computed(() => isDesktop() ? 10 : 3)
+const limit = computed(() => process.client && isDesktop() ? 10 : 3)
 await fetchKeyblockMicroblocks({
   limit: limit.value,
   keyblockHash: route.params.id,
 })
-
 const pageIndex = ref(1)
-const loadPrevMicroblocks = () => fetchKeyblockMicroblocks({ queryParameters: keyblockMicroblocks.value.prev })
-const loadNextMicroblocks = () => fetchKeyblockMicroblocks({ queryParameters: keyblockMicroblocks.value.next })
-// todo shorten naming
+const loadPrevMicroblocks = () =>
+  fetchKeyblockMicroblocks(
+    { queryParameters: microblocks.value.prev })
+const loadNextMicroblocks = () =>
+  fetchKeyblockMicroblocks(
+    { queryParameters: microblocks.value.next })
 </script>
 
 <style scoped>
