+++ src/pages/keyblocks/[id].vue
@@ -11,21 +11,18 @@
     v-if="keyblockDetails"
     :keyblock-details="keyblockDetails"/>
 
-  <keyblock-microblocks-panel :microblocks-count="keyblockDetails.micro_blocks_count"/>
+  <keyblock-microblocks-panel/>
 </template>
 
 <script setup>
 import { storeToRefs } from 'pinia'
 import { useRoute } from 'nuxt/app'
-import { useKeyblockDetailsStore } from '@/stores/keyblockDetailsStore'
 import PageHeader from '@/components/PageHeader'
 import KeyblockDetailsPanel from '@/components/KeyblockDetailsPanel'
-import KeyblockMicroblocksPanel from '~/pages/keyblocks/KeyblockMicroblocksPanel'
-import { useKeyblockDetailsStore } from '@/stores/keyblockDetails'
-import PageHeader from '@/components/PageHeader'
-import KeyblockDetailsPanel from '@/components/KeyblockDetailsPanel'
+import KeyblockMicroblocksPanel from '@/pages/keyblocks/KeyblockMicroblocksPanel'
+import { useKeyblockDetails } from '@/stores/keyblockDetails'
 
-const keyblockDetailsStore = useKeyblockDetailsStore()
+const keyblockDetailsStore = useKeyblockDetails()
 const { keyblockDetails } = storeToRefs(keyblockDetailsStore)
 const { fetchKeyblock } = keyblockDetailsStore
 const route = useRoute()
