+++ src/stores/keyblockDetails.js
@@ -1,16 +1,20 @@
 import { defineStore } from 'pinia'
 import axios from 'axios'
 import { useRuntimeConfig } from 'nuxt/app'
-import { adaptKeyblock } from '@/utils/adapters'
+import { adaptKeyblock, adaptKeyblockMicroblocks } from '@/utils/adapters'
 
-export const useKeyblockDetailsStore = defineStore('keyblockDetails', () => {
+export const useKeyblockDetails = defineStore('keyblockDetails', () => {
   const { MIDDLEWARE_URL } = useRuntimeConfig().public
 
   const rawKeyblock = ref(null)
   const keyblockDeltaStats = ref(null)
+  const rawKeyblockMicroblocks = ref(null)
 
   const keyblockDetails = computed(() => {
     return rawKeyblock.value ? adaptKeyblock(rawKeyblock.value, keyblockDeltaStats.value) : null
+  })
+  const keyblockMicroblocks = computed(() => {
+    return rawKeyblockMicroblocks.value ? adaptKeyblockMicroblocks(rawKeyblockMicroblocks.value) : null
   })
 
   async function fetchKeyblock(keyblockHash) {
@@ -24,6 +28,13 @@ export const useKeyblockDetailsStore = defineStore('keyblockDetails', () => {
     rawKeyblock.value = data
   }
 
+  async function fetchKeyblockMicroblocks({ queryParameters, limit, keyblockHash } = {}) {
+    rawKeyblockMicroblocks.value = null
+    const defaultParameters = `/v2/key-blocks/${keyblockHash}/micro-blocks?limit=${limit ?? 10}`
+    const { data } = await axios.get(`${MIDDLEWARE_URL}${queryParameters || defaultParameters}`)
+    rawKeyblockMicroblocks.value = data
+  }
+
   async function fetchKeyblockDeltaStats(keyblockHeight) {
     keyblockDeltaStats.value = null
     const { data } = await axios.get(`${MIDDLEWARE_URL}/v2/deltastats?scope=gen:${keyblockHeight}`)
@@ -35,5 +46,7 @@ export const useKeyblockDetailsStore = defineStore('keyblockDetails', () => {
     keyblockDeltaStats,
     keyblockDetails,
     fetchKeyblock,
+    fetchKeyblockMicroblocks,
+    keyblockMicroblocks,
   }
 })
